+++ backend/libraries/types/src/diamond_membership.rs
@@ -7,8 +7,6 @@ use std::fmt::{Display, Formatter};
 pub struct DiamondMembershipDetails {
     pub expires_at: TimestampMillis,
     pub pay_in_chat: bool,
-    #[deprecated]
-    pub recurring: Option<DiamondMembershipSubscription>,
     pub subscription: DiamondMembershipSubscription,
 }