+++ frontend/openchat-agent/src/services/storageBucket/candid/types.d.ts
@@ -471,7 +471,6 @@ export interface DeletedContent {
 export interface DiamondMembershipDetails {
   'pay_in_chat' : boolean,
   'subscription' : DiamondMembershipSubscription,
-  'recurring' : [] | [DiamondMembershipSubscription],
   'expires_at' : TimestampMillis,
 }
 export type DiamondMembershipPlanDuration = { 'OneYear' : null } |