+++ frontend/openchat-agent/src/services/user/candid/types.d.ts
@@ -681,7 +681,6 @@ export type DeletedMessageResponse = { 'MessageNotFound' : null } |
 export interface DiamondMembershipDetails {
   'pay_in_chat' : boolean,
   'subscription' : DiamondMembershipSubscription,
-  'recurring' : [] | [DiamondMembershipSubscription],
   'expires_at' : TimestampMillis,
 }
 export type DiamondMembershipPlanDuration = { 'OneYear' : null } |